requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
astor==0.8.1
async-timeout==4.0.3
attrs==23.2.0
babel==2.15.0
bce-python-sdk==0.9.7
blinker==1.8.2
boltons==24.0.0
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.8.2
config==0.5.1
contourpy==1.2.1
cycler==0.12.1
cython==3.0.10
datasets==2.19.1
decorator==5.1.1
dill==0.3.4
distance==0.1.3
distlib==0.3.8
dnspython==2.6.1
docopt==0.6.2
email-validator==2.1.1
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.3
filelock==3.14.0
flake8==7.0.0
flask==3.0.3
flask-babel==2.0.0
flatbuffers==24.3.25
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
future==1.0.0
g2p-en==2.1.0
g2pm==0.1.2.5
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.0
humanfriendly==10.0
identify==2.5.36
idna==3.7
importlib-resources==6.4.0
inflect==7.2.1
iniconfig==2.0.0
itsdangerous==2.2.0
jieba==0.42.1
jinja2==3.1.3
joblib==1.4.2
kiwisolver==1.4.5
lark==1.1.9
lark-parser==0.12.0
latex2mathml==3.75.5
lxml==5.1.0
markdown-it-py==3.0.0
markdown2==2.4.13
markupsafe==2.1.5
matplotlib==3.8.4
mccabe==0.7.0
mdurl==0.1.2
mock==5.1.0
more-itertools==10.2.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.12.2
nltk==3.8.1
nodeenv==1.8.0
numpy==1.26.4
onnxruntime-gpu==1.16.3
opt-einsum==3.3.0
orjson==3.10.3
packaging==24.0
paddle-bfloat==0.1.7
paddle2onnx==1.1.0
paddlefsl==1.1.0
paddlenlp==2.5.2
paddlepaddle==2.3.2
paddlesde==0.2.5
paddlespeech-ctcdecoders==0.2.0
paddlespeech-feat==0.1.0
pandas==2.2.2
pillow==10.3.0
platformdirs==4.2.1
pluggy==1.5.0
pre-commit==3.5.0
protobuf==3.20.0
py-asciimath==0.3.0
pyarrow==16.0.0
pyarrow-hotfix==0.6
pycodestyle==2.11.1
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.1
pydantic-core==2.18.2
pyflakes==3.2.0
pygments==2.18.0
pynini==2.1.5
pyparsing==3.1.2
pypinyin==0.44.0
pypinyin-dict==0.7.0
pytest==8.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
pyyaml==6.0.1
regex==2024.4.28
requests==2.31.0
rich==13.7.1
scikit-learn==1.4.2
scipy==1.13.0
sentencepiece==0.2.0
seqeval==1.2.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soundfile==0.12.1
starlette==0.37.2
sympy==1.12
threadpoolctl==3.5.0
tomli==2.0.1
tqdm==4.66.4
trampoline==0.1.2
typeguard==4.2.1
typer==0.12.3
typing-extensions==4.11.0
tzdata==2024.1
ujson==5.9.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
virtualenv==20.26.1
visualdl==2.4.2
watchfiles==0.21.0
webencodings==0.5.1
websockets==12.0
werkzeug==3.0.3
xxhash==3.4.1
yacs==0.1.8
yarl==1.9.4
pydub
tritonclient[all]
